GUIScript: oops, we need these casts

+++ gemrb/plugins/GUIScript/GUIScript.cpp
@@ -5709,9 +5709,9 @@ static PyObject* GemRB_GetSlotType(PyObject * /*self*/, PyObject* args)
 	}
 
 	PyDict_SetItemString(dict, "Slot", DecRef(PyInt_FromLong, tmp));
-	PyDict_SetItemString(dict, "Type", DecRef(PyInt_FromLong, core->QuerySlotType(tmp)));
-	PyDict_SetItemString(dict, "ID", DecRef(PyInt_FromLong, core->QuerySlotID(tmp)));
-	PyDict_SetItemString(dict, "Tip", DecRef(PyInt_FromLong, core->QuerySlottip(tmp)));
+	PyDict_SetItemString(dict, "Type", DecRef(PyInt_FromLong, (int)core->QuerySlotType(tmp)));
+	PyDict_SetItemString(dict, "ID", DecRef(PyInt_FromLong, (int)core->QuerySlotID(tmp)));
+	PyDict_SetItemString(dict, "Tip", DecRef(PyInt_FromLong, (int)core->QuerySlottip(tmp)));
 	//see if the actor shouldn't have some slots displayed
 	if (!actor || !actor->PCStats) {
 		goto has_slot;
